Add index com e sem interface
